refactor: add common css for forms
