--- frontend/src/components/FormChangePassword/FormChangePassword.module.css
@@ -1,15 +1,0 @@
-.card {
-	width: 20rem;
-}
-
-.form {
-	margin: 2.75rem 0;
-}
-
-.field {
-	margin-bottom: 2rem;
-}
-
-.paddings {
-	padding: 0 1.25rem 2.25rem;
-}
