--- frontend/src/components/FormEditUser/FormEditUser.module.css
@@ -1,25 +1,0 @@
-.card {
-	width: 20rem;
-}
-
-.form {
-	margin: 2.75rem 0;
-}
-
-.field {
-	margin-bottom: 2rem;
-}
-
-.paddings {
-	padding: 0 1.25rem 2.25rem;
-}
-
-.dangerZone {
-	margin-top: 2.75rem;
-}
-
-.dangerBtn,
-.dangerInput,
-.dangerInput input {
-	width: 100%;
-}
