--- frontend/src/components/FormNewEvent/FormNewEvent.module.css
@@ -1,43 +1,0 @@
-.card {
-	width: 20rem;
-}
-
-.form {
-	margin: 2.75rem 0;
-}
-
-.field {
-	margin-bottom: 2rem;
-}
-
-.calendar {
-	max-width: 260px;
-}
-
-.paddings {
-	padding: 0 1.25rem 2.25rem;
-}
-
-.item {
-	display: flex;
-	flex-wrap: nowrap;
-	align-items: center;
-	max-width: 268px;
-}
-
-.optionText {
-	margin: 0;
-	white-space: break-spaces;
-
-	/* overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis; */
-}
-
-.marker {
-	width: 16px;
-	height: 16px;
-	border-radius: 50%;
-	margin-right: 8px;
-	flex-shrink: 0;
-}
