+++ frontend/src/components/Forms/FormEditCalendar.jsx
@@ -6,7 +6,7 @@ import { Button } from 'primereact/button';
 import { RadioButton } from 'primereact/radiobutton';
 import { classNames as cn } from 'primereact/utils';
 import { CurrentUserContext } from '../../context';
-import styles from './FormEditCalendar.module.css';
+import styles from './Forms.module.css';
 import { Color } from '../../utils/common';
 
 export function FormEditCalendar({
