+++ frontend/src/components/Forms/FormRegistration.jsx
@@ -7,7 +7,7 @@ import { Password } from 'primereact/password';
 import { Dialog } from 'primereact/dialog';
 import { Divider } from 'primereact/divider';
 import { classNames as cn } from 'primereact/utils';
-import styles from './FormRegistration.module.css';
+import styles from './Forms.module.css';
 
 export function FormRegistration({
 	showFormLogin,
