+++ frontend/src/components/Forms/index.js
@@ -1,0 +1,17 @@
+import { FormChangePassword } from './FormChangePassword';
+import { FormEditCalendar } from './FormEditCalendar';
+import { FormEditUser } from './FormEditUser';
+import { FormLogin } from './FormLogin';
+import { FormNewCalendar } from './FormNewCalendar';
+import { FormNewEvent } from './FormNewEvent';
+import { FormRegistration } from './FormRegistration';
+
+export {
+	FormChangePassword,
+	FormEditCalendar,
+	FormEditUser,
+	FormLogin,
+	FormNewCalendar,
+	FormNewEvent,
+	FormRegistration,
+};
