+++ frontend/src/components/PopupEditCalendar/PopupEditCalendar.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Dialog } from 'primereact/dialog';
-import { FormEditCalendar } from '../FormEditCalendar/FormEditCalendar';
+import { FormEditCalendar } from '../Forms';
 
 export function PopupEditCalendar({
 	visible,
